refactor: shifted stringToBoolean function to utils and created tests
