+++ src/utils/stringToBoolean.js
@@ -1,0 +1,12 @@
+function stringToBoolean(string) {
+  if (string) {
+    switch (string.toLowerCase().trim()) {
+      case 'true': case 'yes': case '1': return true;
+      case 'false': case 'no': case '0': case null: return false;
+      default: return Boolean(string);
+    }
+  }
+  else return false;
+}
+
+export default stringToBoolean;
